kiwi/requirements.txt
appdirs>=1.4.0
pyyaml>=3.11
colorama>=0.3.7
flask>=0.12.2
eventlet>=0.21.0
jinja2>=2.7.3